[EASY] `import typeof * as`
Summary:
Adds support for `import typeof * as`, which allows one to extract the type of the [ModuleNamespaceObject](http://www.ecma-international.org/ecma-262/6.0/index.html#sec-module-namespace-objects) for a given module. We didn't have this before because the use-case is pretty rare, but it is a legit one.

For example, if the intended purpose of a module is to be used exclusively via it's namespace object (i.e. the NS object is passed around or whatever), it's nice to be able to grab the type of that object directly rather than import + typeof-alias.

Reviewed By: gabelevi

Differential Revision: D2730916

fb-gh-sync-id: e0f1a058a22ab31d47f6faa5b944e53bfd7e2b9e