check Function.prototype properties on AnyFunT
Summary:
instead of `AnyFunT` being a callable version of `AnyObjT`, in which none of its properties are tracked or checked, this diff makes it verify the props from `Function.prototype` but otherwise allow any other property (as it did before).

a parallel change for `AnyObjT` would be a good idea, so you can't do `({}: Object).toString = false`.

Reviewed By: bhosmer

Differential Revision: D2738044

fb-gh-sync-id: 961e33f68c96fb44a44cc7e1b7b99da60249079b