dep calc speedup
Summary:
Improve the incremental dependency calc by using a reverse dependency graph
to generate a closure starting from touched files, rather than using forward deps
to identify a subset of untouched files. In particular this eliminates the substantial
part of the total time that scaled with the number of unmodified files.

It shouldn't be hard to generate the reverse graph once and update it on rechecks,
but leaving that for later.

Also, avoid re-resolving files whose resolution paths aren't affected by incremental
changes. Resolution is expensive because it hits the file system, so this is a
substantial speedup when large numbers of files are touched on a recheck.

Reviewed By: avikchaudhuri

Differential Revision: D2727177

fb-gh-sync-id: 8135b368a92f180827097d4bc9a7c3d4f1d9f6c2