autocomplete failure mode
Summary:
Currently we return empty results when autocomplete fails, without any
indication of what went wrong (if anything).

This leads to a poor user experience, because the user doesn't know whether she
or Flow is to blame (or what she could do to get better results).

This diff changes the (JSON) output format for autocomplete to return an
optional failure message. Editors that use the autocomplete service can now
report this message on failure.

The design of the format is intentionally simple to begin with. Old behavior can
be recovered by looking at the `result` field of the output. When the `error`
field exists, it is a string.

Ideally the same format would be used in other commands. At that point it might
benefit from some further abstraction / modularization: each command result type
could be a module with a common signature, the type of the failure message could
be customized, etc. We leave all that as future work.

Reviewed By: mroch

Differential Revision: D2635549

fb-gh-sync-id: 07da03e7a1c396efa437dbe245736440c29e13c2