this type
Summary:
The current[*] diff introduces `this` types for classes. Support for `this` types
elsewhere is left as future work, which may or may not be done as part of the
current diff.

For supported classes, `this` is the type of `this`, and is a subtype of the
enclosing class. The main benefit is that when viewed through a subclass of the
enclosing class, `this` takes on the meaning of the subclass instead of the
enclosing class, so more good examples typecheck.

However, there are many gotchas. `this` must only appear in "output" positions in
the class signature, otherwise even bad examples may typecheck. And for
unannotated class properties, `this` cannot be inferred (for the same reasons we
cannot infer generics: the connection will become clear below). Lots of examples
and counterexamples are included in the current diff to illustrate these points.

How is `this` implemented? The basic idea is to make every supported class C have
type ThisClassT C instead of ClassT C. ThisClassT C is somewhat like PolyT<`this`>
C, where C might have free occurrences of `this`: these occurrences are bound by
ThisClassT.

Similarly, a polymorphic class C<X> has type Poly<X> ThisClassT C. Note that
`this` is bound only once other (normal) type parameters are bound.

Corresponding notions of `this`-application and `this`-specialization, mirroring
TypeAppT and SpecializeT, are also introduced.

Finally, we implement polarity checking for supported classes, which can be
reused to check implementations of Promise, Generator, etc. and more generally
complete the implementation of type parameter variances.

[*] replacing all occurrences of `this` with `the current`

Reviewed By: bhosmer

Differential Revision: D2747512

fb-gh-sync-id: d3ed7580d3d4f2b9f473bd7bcf7856113befa9b8