demote singleton LB conversion rules
Summary:
Because we use UnionT to represent enums of literals, we're vulnerable
to pathological behavior when big enums are checked against each other
or themselves. The recent addition of an opportunistic membership check
(which fires when an incoming LB is exactly equal to a member of a
UnionT UB) was prevented from firing in many such cases because
unions of literals are represented by collections of Singleton
types, which we currently autoconvert to the corresponding
constrained primitive type when they appear as lower bounds.

As a short term fix, this diff demotes the Singleton conversion
rules by moving them below the UnionT rules in the flow pattern
match. This gives the list membership trap a chance to fire
on the unconverted Singleton ~> UnionT of Singletons type pair.

Longer term, we need to do two things:

Reviewed By: avikchaudhuri

Differential Revision: D2764128

fb-gh-sync-id: ef86f11a9118a51049ff8887664ec9125001ffa7