class expr fix
Summary:
Typechecking of class expressions regressed due to the bound of the `this` type
inside class expressions not being updated properly. Current diff fixes. Also
rearranged some code so that the bound, which is simply a self-reference to the
enclosing class, is locally managed by the function that creates the class type.

Fixes #1191

Reviewed By: bhosmer

Differential Revision: D2773268

fb-gh-sync-id: e741ccf338a45e1a9058bea5849424d00775bbb5