autocomplete fix for type params (including this type)
Summary:
Autocomplete never worked for type parameters: since the same AST was traversed
multiple times with different substitutions for the type parameters (a.k.a.
"generated tests"), types recorded for expressions were sensitive to what the
case was in the last traversal. (Un)fortunately, the recent modeling of the type
of `this` as a type parameter exposed that bug by causing a regression in
autocomplete on `this`.

The current diff fixes the general problem by generating tests for type
parameters in an order that guarantees that for any part of the AST, the last
traversal corresponds to the case where the type parameters are all substituted
by their bounds: so that any expressions in that traversal are guaranteed to get
expected autocomplete results.

Fixes #1197

Reviewed By: mroch

Differential Revision: D2778331

fb-gh-sync-id: 1dd6807ddfd46d236e6a7c429ce431fadbb6361d