[hack] Send "exit code" message to monitor_log instead of typechecker logs
Summary:
Previously, I sent it to the typechecker logs because no monitor logs
existed. However, as we add more logic to the monitor, it is clear that this is
a fairly janky solution, and now that we have monitor_log it is redundant. This
change allows us to remove log_mode from the server_process record.

Broader goal: There is good evidence that the number of TYPECHECKER_EXIT Scuba
events that get recorded is much lower than the real number of bad exits. I'd
like to add a timestamp to the "exit code" messages so that I can use
hypershell to find out e.g. how many bad exits there were in a given day. The
changes in this diff make it easier to use Hh_logger for this purpose.

Reviewed By: achow

Differential Revision: D2776811

fb-gh-sync-id: ca7fc62ee375916974a888182f684bbdfdf86d5f