[hack] Encapsulate NamingGlobal.env as NamingGlobal.GEnv
Summary:
`NamingGlobal.env` is a record, but we don't want users of this type to rely on
this implementation detail. I encapsulate this record as a type within a module
and the fields of the record are accessed via methods defined in the module.

Most of this was straight forward with the only sizable change was changing how
the autocomplete service worked. It did a bunch of ad-hoc operations which I changed
to use `NamingGlobal.make_env` and made sure we added the newly declared elements to
the naming heap. Also since we can't use the naming environement any more to get the
name of all classes and functions I had to compute it from the files_info

Reviewed By: int3

Differential Revision: D2775186

fb-gh-sync-id: b75dcdeaeb46f80cdfcac7777c7ef7122242e5a0