[hack] Pass FileInfo.fast via argument instead of global storage for Typing_check_service
Summary: We store the mapping of filenames to names of entities contained to it in `GlobalStorage`. Since this grows with the number of files it makes more sense to pass it as part of the argument for a worker job. This way each worker only receives the data it needs to complete its job.

Reviewed By: int3

Differential Revision: D2778729

fb-gh-sync-id: 5a47aab2abd651d8c336c9496045506f4c8fe785