[easy] formally split def and use types
Summary:
So we've had this informal distinction between def types and use types: the
former describe values, the latter describe operations. In code, we enforce the
distinction via asserts in a few places: e.g., lower bounds are never use types,
so the left argument of `flow` and both arguments of `unify` cannot be use
types. On the other hand, a lot of functions only make sense on def types (e.g.,
substitution), but are doomed to be partial functions as long as we cannot
isolate def types.

This diff formally splits def types into their own datatype, and does the
necessary refactoring. This enables clean solutions to both of the above
problems. Moreover, while old kinks in the model become more obvious, some new
kinks are uncovered in the course of refactoring. But we avoid fixing any of
those kinks in this diff, in the interests of minimizing rebase pain: they
should be addressed in follow-up diffs.

Reviewed By: jeffmo

Differential Revision: D2789023

fb-gh-sync-id: a5480c24aafab0f548a0fd9bed5151d1ddfa720d