[hack] ServerMonitor can monitor multiple processes
Summary:
Instead of taking a single process starter, it takes a list of them. After establishing the connection and checking build ID, client sends the name of the process he wants to talk to.

If any of the monitored processes dies (intentionally or unexpectedly), monitor exits too. When monitor exits, it kills all the processes (so the entire system is healthy, or entire system dies - not foreseeing much need or possibility to recover from partial failure).

Reviewed By: alexchow

Differential Revision: D2782728

fb-gh-sync-id: 03d31c4a9e36cd0b113deedfbc08daaad067013a