changes to switch env handling
Summary:
Changes to how we manage environment state while traversing switch statements.
Main benefit is properly recognizing when a switch is exhaustive, which fixes
false positive "possibly uninitialized" errors in cases where exhaustiveness
guarantees initialization of a var or let.

Note: this is simple exhaustiveness, guaranteed by a default case. The more
subtle problem of recognizing exhaustiveness in the absence of a default -
when all values of the discriminant's type have a corresponding case -
is still outstanding.

Added tests.

Reviewed By: avikchaudhuri

Differential Revision: D2788208

fb-gh-sync-id: b9507a3ed29c24e726c73e95c58b47ede93a7fda