def site variance
Summary:
Add syntax for declaring variance of type parameters. This is enough to provide
general support for def-site variance, because:

(1) Subtyping of generics was already sensitive to variance to accommodate
Promise/Generator use-cases

(2) Variance checking was already implemented for this-types

The core lib definitions of Promise/Generator are now updated to declare
variances using syntax (instead of relying on hard-coded special handling of
those definitions in the typechecker).

Reviewed By: jeffmo

Differential Revision: D2773424

fb-gh-sync-id: 59a1bb501f638b0c304c7250d99144d0bb029137