Refine boolean sentinel properties
Summary:
Adds support for comparing sentinel props against bools.

The challenge here is that we support equality refinements on bools that we don't (yet) support on strings or numbers, making this trickier than the others. That is, `if (foo.bar === false)` already refines `foo.bar` to be `false`, but we don't currently also refine `foo` based on its sentinel prop.

At a high level, all this diff does is make `bool_test` call `sentinel_prop_test` if the expression being compared is a member expression, in addition to the normal refinement it was already doing.

At a low level, this diff is attempting to avoid walking the AST twice (once in `sentinel_prop_test` to extract the object and prop separately, and again in `bool_test` since it didn't previously care about the structure of the operand). This refactoring also sets us up to fairly easily add refinements based on string and number equality without running into similar issues with sentinel props.

Fixes https://github.com/facebook/flow/issues/577

Reviewed By: bhosmer

Differential Revision: D2805165

fb-gh-sync-id: 31949266a35b2e551fee8f3d6776d8e60ae14b3a