[cleanup] fix all ocaml warning 45's
Summary:
we were suppressing warning 45, "Open statement shadows an already defined label or constructor," which let us play fast and loose with labels and constructors that had the same name in different modules.

this is a valuable warning because it's just as confusing to a human as it is to the compiler when hard-to-see `open`'s change the meaning of things.

Reviewed By: gabelevi

Differential Revision: D2814380

fb-gh-sync-id: 56737e5507d6ce1a9e8bd6e51515c1bb4eac0c72