handle `this` in interface
Summary:
We were'nt handling `this` consistently in interfaces: on the one hand, we were
allowing it and processing it in interface definitions, but on the other hand,
we're assuming that interfaces were just like types, with `this` "closed" (in
the fixpoint sense): not only when they're used (say) to make types but also
when they're used to extend other interfaces.

After a bit of thought, I'm deciding to disallow `this` in interface
definitions. This is not a restriction unless we build hierarchies of
interfaces, which we've not seen happen in JS (classes implicitly matching
interfaces without needing to declare them may have a lot to do with
it). There's some benefit to relaxing this decision in the future, but we have
to realize that unlike other languages, interfaces don't just describe classes
in JS: they describe arbitrary objects, and arbitrary objects don't (yet) bind
self the same way as we pretend classes do (instead, they are bags of methods,
and each method binds its own self). So for now, `this` remains a "classy"
concept, pending further work on specifying self types for methods.

Reviewed By: gabelevi

Differential Revision: D2836691

fb-gh-sync-id: 47af7c29f0d363cc5df265ab4efd788bfc805be6