super() tests
Summary:
added tests for `super()`, especially around a superclass returning `this` from its constructor.

`super()` is essentially `this.prototype.constructor.call(this)`. we were doing `this.prototype.constructor.call(this.prototype)`. it doesn't seem to actually matter, though, because of the way we handle `this`.

if the parent is annotated `constructor(): this { return this; }` then everything works fine. without the return annotation, if a child does `constructor() { return super(); }`, flow incorrectly thinks that `new Child()` returns a `Parent` instance. not sure if this is a problem, or intentional because we don't infer generics.

Reviewed By: gabelevi, avikchaudhuri

Differential Revision: D2810588

fb-gh-sync-id: 6c4d45f87c8d5bca91563d62d115c5b9189ed184