[hack] Remove all_classes from typing_decl
Summary:
This work was prompted by the goal of making lazy decl possible. For that to
happen, the type-check phase must have access to the same global state that
type-decl has. In particular, it is missing the all_classes map of class names
to containing files. However, I realized that this state is already globally
available thanks to dlreeves' NamingGlobal work. The ClassPosHeap contains this
mapping (as well as including the position of the identifier).

There is one semantic change involved, though I do not think it is material.
First, note that all_classes records a mapping of class names to _all_ the
files they were declared in, meaning that duplicate declarations would be
recorded. ClassPosHeap only keeps the first one it sees. However, I do not think
keeping duplicates is very useful -- we only need that map to ensure that if
Bar extends Foo, then (some) Foo is declared before we try and inline its
members into Bar. Regardless, this is a moot point because if a class has a
duplicate declaration, then it is an error in the naming phase, and we
remove[1] these errors from the fast (and therefore the all_classes map as
well).

There is a comment in namingGlobal about dealing carefully with duplicates[2].
For good measure, I added an integration test for the case where we have
multiple parents for a given child class. This works fine, the only difference
being that instead of considering the child class to have unknown parents and
therefore crippled typechecking of field members, we pick one of the
duplicates.

Performance
==========

As might be expected, this further speeds up type-decl. On my devserver it goes
from 26-28s down to 20-22s, about a 22% improvement. Other phases are
unaffected.

[1]: https://github.com/facebook/hhvm/blob/8885622b1c0395d437ad98429fa5be23fc8f3125/hphp/hack/src/server/serverInit.ml#L360

[2]: https://github.com/facebook/hhvm/blob/8885622b1c0395d437ad98429fa5be23fc8f3125/hphp/hack/src/naming/namingGlobal.ml#L175

Reviewed By: dlreeves

Differential Revision: D2848357

fb-gh-sync-id: 515427c21f9c3fc046287903cc0e674b453abe83