support sentinel prop existence refinements
Summary:
allows refinement of a union based on the truthiness of a sentinel property. for example:

```
type X = { done: true, value: string } | { done: false };
function test(x: X): string {
  if (x.done) {
    return x.value;
  }
  return "not done";
}
```

(this example is similar to boolean sentinel prop tests, `if (x.done === true)`)

like all other sentinel prop tests, but perhaps more confusingly in this case, we have to assume that the property *may* exist if it is not explictly falsy, due to subtyping. so, `if (x.prop)` can filter out unreachable branches of the union if they are `{ prop: false }` or `{ prop: "" }` or `{ prop: 0 }`, but will not filter out a branch that doesn't contain a `prop` property at all.

Reviewed By: bhosmer

Differential Revision: D2822755

fb-gh-sync-id: eda18f25332d3fbed22d24f39d0d1a6e052fecbf