[hack] Add style guide
Summary:
Let's get the codebase into a more consistent state. This file will also be
helpful for new / external contributors.

The guide is intended as a proposal; please let me know if you find any of it
objectionable.

I'd like buy-in from everyone, so I'll land this after all reviewers
have stamped it.

Reviewed By: dlreeves, dabek, alexchow

Differential Revision: D2815191

fb-gh-sync-id: c1693bf2cca43cc43ff9c6b3a55d7c49ce655ead