[easy] move master_cx creation out of Flow_js
Summary: the master context is created once, in the master process. It is currently passed into `merge_component_strict` and `restore`, but not the new `merge_lib_file`, so this brings parallelism there. Then nothing in `Flow_js` needs it, so I moved it into the code that already runs in the master, so we can't possibly accidentally call it from the workers.

Reviewed By: bhosmer

Differential Revision: D2863655

fb-gh-sync-id: 4911f1e8573489cc8256797de2cff5accebe0218