[hack] Move typedef constraint checking to typecheck phase
Summary:
Previously, it was being done in the type decl phase, which was likely racy.
The constraint was not guaranteed to be declared before the concrete type, so
when sub_type called get_typedef, it was not guaranteed to succeed.

Indeed, we did see some typedef-related failures in our integration stress
tests (along the lines of "classname is not compatible with string").

Additionally, a clear separation of decl-phase and typecheck-phase heap access
is essential for the implementation of lazy decl.

Reviewed By: dabek

Differential Revision: D2863831

fb-gh-sync-id: 8803308b775ca57e09896f4ef9453b28019f87f5