Fix parsing perf regression
Summary:
bhosmer noted that {D2792581} caused a perf regression in parsing.
After investigating I noticed that Flow really was parsing twice as many files
(avikchaudhuri was right) though it merges them into a set afterwards, so it doesn't
infer twice as many.

The cause of the diff was that the new find code doesn't put `/` at the end of
directories. This broke some code that checked to see if a directory was a
symlink. It would incorrectly conclude that every directory contained a symlink
and then would start over again with a better path filter.

Reviewed By: bhosmer

Differential Revision: D2873888

fb-gh-sync-id: 287b59f65b20ccfe659f6edb87aa41a10b4d35bc