[hack] IDE server process and IDE client mode to connect to it - transport layer
Summary:
Building on top of D2782728, we want to have a separate IDE process dedicated to answering IDE queries, which will not be blocked on heavyweight typechecker work. The purpose of "hh ide" client mode is to establish a persistent connection to this process, and forward all the commands from stdin to it.

Transport layer protocol:

All messages are text, composed of header and body. Header contains the length (in bytes/characters) of the body. Body is an arbitrary text of given length, though at application layer (when we implement it) it will be only JSON encoded objects. For now it accepts everything and always sends back placeholder message - don't pay too much attention to it's format and contents.

Header must be newline terminated. Body can be newline terminated (in which case newline counts towards message length), but doesn't have to. Body can be multiline too.

All the transport layer errors are communicating by client exiting with specific error code. After successful connection all the communication happens by exchanging messages.

Reviewed By: alexchow

Differential Revision: D2827262

fb-gh-sync-id: e32de5ed226863a41bbd8cd5fb985fa5258cab42