[hack] Restart all processes when restarting any of them
Summary:
D2826365 added possibility of retry logic. It does so on per-process basic, but in a scenario of multiple cooperating processes it will be pretty hard for one of them to recover when the other is restarted in random moment - pipes will get broken, shared memory in undefined state, etc.

Instead of this, we kill all the processes and re-do the whole initialization process (that currently doesn't do anything, but in the future will involve creating pipes between processes).

Reviewed By: int3, alexchow

Differential Revision: D2862817

fb-gh-sync-id: dba7295952210339892fecbfb796f6f68a1a6541