Infer this.props: Object for React.createClass() w/o propTypes
Summary:
If you create a react component with createClass() but don't specify the
propTypes, what should the type of props be?

It used to be an empty object, but we didn't enforce that correctly, so
people could do whatever they wanted with this.props.

As of 0.21.0, it started to be an error when people used this.props in a
strict equality situation. It was weird that this was only sometimes
enforced, so glevi changed this.props to be Object by default.

We may change this back to the empty object at some point and fix the
situations where it didn't used to error

Reviewed By: jeffmo

Differential Revision: D2886991

fb-gh-sync-id: e1b6ef25eb2ce378a1e50bd9416c1ad116ca18da