eval destructuring
Summary:
As explained in the diff that introduced it, DestructuringT is a wrapper that
evaluates certain type terms lazily to make some other tasks, like substitution,
simpler. Unfortunately, as implemented, the evaluation created a new tvar, to
hold the result, every time it was invoked. This not only caused redundant
computation, but in some cases could be exploited to cause non-termination.

This diff arranges it so that a particular DestructuringT always holds its
result in the same tvar, when it is evaluated. (This is much like how lazy terms
in Haskell get evaluated.) Moreover, this kind of wrapper is general enough that
it merits a separate classification, so this diff does a bit in that direction,
beyond just fixing the immediate non-termination bug. (As noted elsewhere,
another example of this kind of wrapper is TypeAppT, but moving it over to the
new classification is left as future work.)

Reviewed By: bhosmer

Differential Revision: D2885163

fb-gh-sync-id: 9354c1bc4c46d24a3f8d0a276680c644b739e181