[cleanup] add some flowconfig options to Options
Summary:
`Options` represents the union of CLI and `.flowconfig` options used to start the server. In some cases, we have CLI flags that override the flowconfig, and I think we should allow this for pretty much every flowconfig option. This would make testing easier, for example.

Passing `Options` around more also removes a couple places where we pull the flowconfig out of thin air, and reducing globals is good.

Reviewed By: samwgoldman

Differential Revision: D2887666

fb-gh-sync-id: 1dcbc4e2ded78f722f9a98680303a201fd8e8e86