fix mistakes in code examples
Saying that `b` is type `Foo` or "of type" `Foo` doesn't make any sense, all you can say is that `b` contains a property of type `Class<Foo>`.

Moreover these code samples were clearly intended to demonstrate that the type of `new b.f()` is `Foo` and as such can be assigned to a variable of type `Foo` but can't be assigned to a variable of type `Bar`.  However, they didn't actually demonstrate that at all.  The missing prop error caused by `new b.g()` is irrelevant to this explanation of `typeof`.

I've updated the code and comments to illustrate the effect of `typeof` on instantiating with `new` clearly.