[hack] Client wakeup should be handled by monitor
Summary:
While working on the exit-on-NFS diff, I noticed that hh_client would hang
forever if the typechecker process exited. It was waiting on a wakeup message
from the typechecker that would never arrive.

I realized that there was no need for the typechecker process to send the
wakeup message -- it's only necessary before the server socket is ready. Now
that the monitor process handles the socket connection, we can have it wake up
the client as well.

This *does* break `hh_client start --wait`, but I think it's a pretty redundant
flag: if the user wants to wait for hh_client to start up, they can simply call
`hh_client check`. I've replaced it with a deprecation message.

Reviewed By: alexchow

Differential Revision: D2884554

fb-gh-sync-id: 7fcb4e54edba131394ca527e6b96472bcbb97a51