Fix temp_dir test when run inside a deep directory tree
Summary:
socket structs have a very limited amount of space for the socket address. the sock file is normally stored in the directory specified by `--temp-dir`, but if the path is too long, flow has to ignore `--temp-dir` and put it in `/tmp`. this breaks the test, which wasn't accounting for that.

this fixes the test by using a guaranteed-short `--temp-dir`, so it doesn't depend on where the repo is checked out.

Reviewed By: samwgoldman

Differential Revision: D2903299

fb-gh-sync-id: eaa28be059a6d99da063f366f1ca761749364cec