use CustomFunT to handle Object.assign
Summary:
we model `copyProperties` as if it was synonymous with `Object.assign` (even though it isn't, it copies all enumerable properties not just own-properties). This diff renames our special `CopyProperties` facebookism to `ObjectAssign`, and uses it in `Object.assign` so that the normal flow mechanism handles it instead of pattern matching `Object.assign` in type_inference_js.

one side-effect is that we now have to wait until the lib is merged for `Object.assign` calls to return, which caused two "Missing annotation" errors in the tests to disappear. This is an unrelated bug that should be fixed separately.

Reviewed By: avikchaudhuri

Differential Revision: D2808713

fb-gh-sync-id: c29bc8437094e9d3902676695614e21e56ef9ae7