don't parse flow comments when not in flow mode
Summary: When the parser's `types` mode is off, the parser should act like a normal JS parser as much as possible. We were parsing flow comments like `/*: number */` as a type annotation `: number`, leading to an error like "Type annotations are not allowed in untyped mode." This should just be parsed as a normal comment in untyped mode.

Reviewed By: jeffmo

Differential Revision: D2872076

fb-gh-sync-id: b48d1127bbdbec8c7e8affb089a6ca353b24e417