[hack] Remove get/set_all_locals entirely
Summary:
This makes it statically guaranteed that the set of all_locals is
monotonically increasing (see also D2909525).

The removal of the get/set_all_locals calls in the try-catch analysis is not a
semantic change because the `branch` function that it wraps does exactly the
same thing -- it saves the set of locals and then restores it afterward.

Reviewed By: dlreeves

Differential Revision: D2909532

fb-gh-sync-id: 073b86e57e3019b05beb9b23fc56fbdc17d9ad0a
shipit-source-id: 073b86e57e3019b05beb9b23fc56fbdc17d9ad0a