[hack] Move ident.ml under heap/
Summary:
Ident depends on hh_shared.c because it uses shared memory to ensure that the
identifier it returns is globally unique. However, utils/ doesn't have an
explicit dependency on heap/ (and we don't want to add one), so any target
wanting to use utils/ must also list a dependency on heap/ in order to avoid
link errors -- as discovered by alexchow. This diff fixes that.

Reviewed By: mroch

Differential Revision: D2910206

fb-gh-sync-id: ea8def0bc16492df04cc31852351bbb483e08ca6
shipit-source-id: ea8def0bc16492df04cc31852351bbb483e08ca6