[hack] Make all_locals an abstract type
Summary:
There's no real good justification for this -- I just thought it
looked better.

The test output changes because in the case of

```
function f(): int {
  switch (1) {
    case 1:
      break;
    case 2:
      switch ('x') {
        case 'x':
          $x = 2;
          break;
        default:
          $x = 4;
          break;
      }
      break;
    case 3:
      $x = 8;
      break;
  }

  return $x;
}
```

we would previously raise an error on `$x = 8`, but now we raise it on `$x = 2`. I don't think it's a material change though -- it's a pretty clear error message in both cases

Reviewed By: alexchow

Differential Revision: D2909540

fb-gh-sync-id: ca76ef8c7f2e164a4ba19350d849fcf8dfc8bc85
shipit-source-id: ca76ef8c7f2e164a4ba19350d849fcf8dfc8bc85